+++ modules/server/src/main/scala/scaladex/server/route/package.scala
@@ -2,9 +2,6 @@ package scaladex.server
 
 import java.time.Instant
 
-import scala.concurrent.ExecutionContext
-import scala.concurrent.Future
-
 import akka.http.scaladsl.server.Directive1
 import akka.http.scaladsl.server.Directives.Segment
 import akka.http.scaladsl.server.Directives._
@@ -12,11 +9,9 @@ import akka.http.scaladsl.server.PathMatcher
 import akka.http.scaladsl.server.PathMatcher1
 import akka.http.scaladsl.unmarshalling.Unmarshaller
 import scaladex.core.model.Artifact
-import scaladex.core.model.ArtifactSelection
 import scaladex.core.model.Project
 import scaladex.core.model.SemanticVersion
 import scaladex.core.model.search.PageParams
-import scaladex.core.service.WebDatabase
 
 package object route {
 
@@ -39,48 +34,4 @@ package object route {
 
   def paging(size: Int): Directive1[PageParams] =
     parameter("page".as[Int].withDefault(1)).map(PageParams(_, size))
-
-  def getSelectedArtifact(
-      database: WebDatabase,
-      org: Project.Organization,
-      repo: Project.Repository,
-      binaryVersion: Option[String],
-      artifact: Option[Artifact.Name],
-      version: Option[String],
-      selected: Option[String]
-  )(implicit ec: ExecutionContext): Future[Option[Artifact]] = {
-    val artifactSelection = ArtifactSelection.parse(
-      binaryVersion = binaryVersion,
-      artifactName = artifact,
-      version = version,
-      selected = selected
-    )
-    val projectRef = Project.Reference(org, repo)
-    for {
-      project <- database.getProject(projectRef)
-      artifacts <- database.getArtifacts(projectRef)
-      defaultArtifact = project
-        .flatMap(p => artifactSelection.defaultArtifact(artifacts, p))
-    } yield defaultArtifact
-  }
-  def getSelectedArtifact(
-      database: WebDatabase,
-      project: Project,
-      binaryVersion: Option[String],
-      artifact: Option[Artifact.Name],
-      version: Option[String],
-      selected: Option[String]
-  )(implicit ec: ExecutionContext): Future[Option[Artifact]] = {
-    val artifactSelection = ArtifactSelection.parse(
-      binaryVersion = binaryVersion,
-      artifactName = artifact,
-      version = version,
-      selected = selected
-    )
-    for {
-      artifacts <- database.getArtifacts(project.reference)
-      defaultArtifact = artifactSelection.defaultArtifact(artifacts, project)
-    } yield defaultArtifact
-  }
-
 }
